Backend/requirements.txt
# ElevenLabs Conversational AI
elevenlabs==1.9.0

# Environment variable management
python-dotenv==1.0.0

# HTML parsing for web scraping
beautifulsoup4==4.12.3
lxml==5.1.0

# HTTP requests (if needed)
requests==2.31.0
